% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiance.R
\name{ambiance}
\alias{ambiance}
\title{Switch ambiance}
\usage{
ambiance(context)
}
\arguments{
\item{context}{Name of an ambiance among those defined with \code{define_ambiances}}
}
\value{
Change RStudio theme
}
\description{
Switch ambiance
}
