% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_ambiances}
\alias{define_ambiances}
\title{Define a list of ambiances}
\usage{
define_ambiances(ambiances)
}
\arguments{
\item{ambiances}{This must be a list of vectors. Each vector should have two components: a name and a theme. See \code{Examples} for more details.}
}
\value{
Creates a CSV file stored where the package is stored.
}
\description{
This function creates a CSV file that contains a list of ambiances. The CSV file is then called by \code{ambiance()} to apply specific themes.
}
\examples{
\dontrun{
define_ambiances(
  list(
    c("perso", "cobalt"),
    c("class", "textmate (default)")
  )
)
}
}
